internal/worker: add query param to serve vulncheck/scan
Add a "serve" query param to the vulncheck/scan endpoint.
A URL with "serve=true" will scan as usual, but write the
result back to the client as JSON instead of uploading it
to BigQuery.

This makes it easier to test scanning without writing any data.

Previously we had the test-vulncheck-sandbox endpoint for this,
but it followed different logic so it wasn't as good a test.

Change-Id: If7037ff4fe3a11a082ac3c03c752a16339031c1e
Reviewed-on: https://go-review.googlesource.com/c/pkgsite-metrics/+/471235
Run-TryBot: Jonathan Amsterdam <jba@google.com>
TryBot-Result: Gopher Robot <gobot@golang.org>
Run-TryBot: Zvonimir Pavlinovic <zpavlinovic@google.com>
Reviewed-by: Zvonimir Pavlinovic <zpavlinovic@google.com>