internal/worker: run "go clean" in sandbox directly
We want to run "go clean" in the sandbox to remove cached files,
so we don't run out of disk (which is memory in Cloud Run).

Previously, we ran the vulncheck_sandbox binary, and it had
code to run go clean.

But we could just invoke "go" in the sandbox and pass it "clean"
and the other arguments.

This CL does that, and removes the now-dead code from
vulncheck_sandbox. It also verifies that go clean is doing something
by logging the disk usage before and after.

Change-Id: Ieac31588d21e4db5f60933ed20ac88516c920eeb
Reviewed-on: https://go-review.googlesource.com/c/pkgsite-metrics/+/471162
Reviewed-by: Maceo Thompson <maceothompson@google.com>
Run-TryBot: Jonathan Amsterdam <jba@google.com>
TryBot-Result: Gopher Robot <gobot@golang.org>