cmd/vulncheck_sandbox: categorize package loading errors
This is achieved indirectly by forcing the load.Packages function to
classify loading errors. This in turn implies some naming changes to the
error constants in derrors package.

The rationale for this approach is that package loading is irrespective
of the overall analysis, so these analyses should classify the errors
the same way.

Change-Id: I00c0dfeea816ccc7e5d8942f14a9b7be75251524
Reviewed-on: https://go-review.googlesource.com/c/pkgsite-metrics/+/473396
Reviewed-by: Jonathan Amsterdam <jba@google.com>
Run-TryBot: Zvonimir Pavlinovic <zpavlinovic@google.com>
TryBot-Result: Gopher Robot <gobot@golang.org>