terraform: add alert for build failures
Add a log-based metric that looks for failures in Cloud Build logs,
and an alert that fires if that metric is positive.

This is not the official way to be notified about Cloud Build
events. You're supposed to listen to the pubsub channel it writes to,
but that involves setting up a server or Cloud Function that then
sends email.

Instead, the metric looks for the line "ERROR" in the logs. I observed
that this line occurs when a build fails. But it might not be totally
reliable; we should keep an eye on it.

Change-Id: I5630fa7d6f6444fc835f1dfe4b94d72a905af6df
Reviewed-on: https://go-review.googlesource.com/c/pkgsite-metrics/+/473168
Reviewed-by: Zvonimir Pavlinovic <zpavlinovic@google.com>
Run-TryBot: Jonathan Amsterdam <jba@google.com>
TryBot-Result: Gopher Robot <gobot@golang.org>