internal/bigquery,vulncheck: only read work version columns
Reading work versions was extremely slow. The reason was that we were
reading the entirety of every row. This CL reads just the columns we
need, giving a huge speedup.

To do this, we needed a way to specify columns in a partition
query. The arguments to the PartitionQuery function were all strings,
so it was already a bit confusing what each string meant, and this was
going to be one more. PartitionQuery is now a struct, so the field
names can be written in the code.

Change-Id: I96205796b0a44220bdd9373332311e210d3176a5
Reviewed-on: https://go-review.googlesource.com/c/pkgsite-metrics/+/474896
TryBot-Result: Gopher Robot <gobot@golang.org>
Run-TryBot: Jonathan Amsterdam <jba@google.com>
Reviewed-by: Zvonimir Pavlinovic <zpavlinovic@google.com>