all: change vulncheck endpoints to govulncheck
Note that this CL still does not change the name of the table. This will
come in a separate CL together with table schema change.

This changes comes as a part of CL series that replace the uses of
"vulncheck" with "govulncheck".

Change-Id: I044b807752e67beb10bf1b10569c3c500188ebb8
Reviewed-on: https://go-review.googlesource.com/c/pkgsite-metrics/+/476495
Run-TryBot: Zvonimir Pavlinovic <zpavlinovic@google.com>
Reviewed-by: Jonathan Amsterdam <jba@google.com>
TryBot-Result: Gopher Robot <gobot@golang.org>