internal/worker: check analysis work version
Before running an analysis scan, check if the result is already
in BigQuery.

As with govulncheck, we do this by storing a WorkVersion for
each analysis, keyed by module path, version, and binary. If the work
version of the incoming work matches a stored one that is read at
startup, then the scan can be skipped.

This required some refactoring to download the binary earlier,
so its hash can be computed before scanning starts.

Change-Id: I24506c7669ed282224c987d0700123820dd64ab4
Reviewed-on: https://go-review.googlesource.com/c/pkgsite-metrics/+/476955
Run-TryBot: Jonathan Amsterdam <jba@google.com>
TryBot-Result: Gopher Robot <gobot@golang.org>
Reviewed-by: Zvonimir Pavlinovic <zpavlinovic@google.com>