cmd/worker: pin the govulncheck version to the one in go.mod file
govulncheck is going through certain changes to which we currently
are not adjusted to. Until it stabilizes, we fix a version.

The transition from this state to the one where govulncheck is used at
the latest (stable) version will happen gradually. Hence, more of CLs
such as this one can be expected.

Change-Id: I6edcf564d17b75cfa5282ea04c0cdc330d0aee68
Reviewed-on: https://go-review.googlesource.com/c/pkgsite-metrics/+/478675
Reviewed-by: Maceo Thompson <maceothompson@google.com>
Run-TryBot: Zvonimir Pavlinovic <zpavlinovic@google.com>
TryBot-Result: Gopher Robot <gobot@golang.org>