internal/worker: create BigQuery tables explicitly
It's a nice feature that the bigquery client remembers the schema for
a table, but it shouldn't be used to blindly create all the registered
tables in the same dataset. Some tables might belong in a different
dataset.

In this CL, have the worker create the tables it needs explicitly
instead of relying on the list in the bigquery client.

In fact, remove the ability to get that list.

Change-Id: I76d018bb3993b6f344c62f5da4d018ef07dd2d33
Reviewed-on: https://go-review.googlesource.com/c/pkgsite-metrics/+/482161
Run-TryBot: Jonathan Amsterdam <jba@google.com>
TryBot-Result: Gopher Robot <gobot@golang.org>
Reviewed-by: Zvonimir Pavlinovic <zpavlinovic@google.com>
Auto-Submit: Jonathan Amsterdam <jba@google.com>