internal/worker: remove govulncheck.WorkVersion.VulnVersion
The VulnVersion field recorded the version of the golang.org/x/vuln
module, so changes to that module would trigger re-scanning.

But the WorkerVersion field, which holds the docker image ID, subsumes
that: a change to the version of x/vuln linked into the worker binary
implies a changed docker image.

Furthermore, now that we run govulncheck in a separate process, the
version of the module linked into the worker is irrelevant.

Change-Id: I332615280b57b0636d53efd2e42926ad120f7026
Reviewed-on: https://go-review.googlesource.com/c/pkgsite-metrics/+/483697
TryBot-Result: Gopher Robot <gobot@golang.org>
Reviewed-by: Zvonimir Pavlinovic <zpavlinovic@google.com>
Run-TryBot: Jonathan Amsterdam <jba@google.com>