all: correct removal of testing datasets
In integration tests, we create test datasets that we try to delete once
tests are done. This does not work since Delete() API we use expects
datasets to be empty. We correct this by using an approprate API.

Also, a test dataset is unique up to the current date (second
precision), but if two tests are executed in parallel at the same
second, this can lead to races. We now create datasets unique to the
tests creating them.

The major impact of this CL is that the target Cloud project will not
accumulate test datasets that are of no use.

Change-Id: Idf3dbbe6f16b9848e8f017c808605dbfebfc7f43
Reviewed-on: https://go-review.googlesource.com/c/pkgsite-metrics/+/500515
TryBot-Result: Gopher Robot <gobot@golang.org>
Reviewed-by: Jonathan Amsterdam <jba@google.com>
Run-TryBot: Zvonimir Pavlinovic <zpavlinovic@google.com>