internal/jobs, etc.: add binary info to job data
Add the binary name, version (hash) and args to the Job struct.

With that information, ejobs can match a job to result rows in
BigQuery. That paves the way for providing an ejobs command that
can download the results for a job.

Change-Id: Ifd1ec24bbe11da80998dd3267e4882a29f3075ed
Reviewed-on: https://go-review.googlesource.com/c/pkgsite-metrics/+/511756
Run-TryBot: Jonathan Amsterdam <jba@google.com>
TryBot-Result: Gopher Robot <gobot@golang.org>
Reviewed-by: Maceo Thompson <maceothompson@google.com>