internal/govulncheck: fix error value for compare finding
Otherwise, unmarshaling fails. The marshaling and unmarshaling are
needed to communicate results from the govulncheck compare sandbox. We
can marshal the error type, but unmarshaling to an interface type will
fail if the underlying type is not one of the supported ones:
https://pkg.go.dev/encoding/json#Unmarshal.

This CL also propagates some errors to the top for easier monitoring.

Change-Id: I72d1123d75cda40959e323cf5d241375fc8321c6
Reviewed-on: https://go-review.googlesource.com/c/pkgsite-metrics/+/523815
Reviewed-by: Jonathan Amsterdam <jba@google.com>
TryBot-Result: Gopher Robot <gobot@golang.org>
Run-TryBot: Zvonimir Pavlinovic <zpavlinovic@google.com>