internal/bigquery: order columns in SchemaVersion
In relational algebra and hence bigquery, the order of columns does not
matter. Hence, two schemas with same columns but different ordering in
code definition should be regarded as same. SchemaVersion did not
account for that before. It does so now.

This problem manifested when patching tables in bq: adding a new field
to the middle of govulncheck.Result would result in appending that
column to schema in bq. The schema equality check would then fail
although it should not.

Change-Id: I1d47f59d1d8a415e2976ebe699c37a7b2180a8f6
Reviewed-on: https://go-review.googlesource.com/c/pkgsite-metrics/+/526556
TryBot-Result: Gopher Robot <gobot@golang.org>
Run-TryBot: Zvonimir Pavlinovic <zpavlinovic@google.com>
Reviewed-by: Jonathan Amsterdam <jba@google.com>