internal/worker: do not propagate CompareModule errors
An endpoint should return an error only when we think this error is
transient, i.e., it might go away with a retry. Otherwise, a task can be
retried indefinitely which might clog the queue.

This is happening for compare mode. We log the error message for
visibility but we do not propagate it back. Note that in govulncheck
mode we save the error to the table, but this does not make sense for
compare mode.

Change-Id: I106d182406e12ffc17b84f9e98d60f3daa00f110
Reviewed-on: https://go-review.googlesource.com/c/pkgsite-metrics/+/543455
Reviewed-by: Jonathan Amsterdam <jba@google.com>
Run-TryBot: Zvonimir Pavlinovic <zpavlinovic@google.com>
TryBot-Result: Gopher Robot <gobot@golang.org>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>