internal/worker: re-analyze vendor issue projects
We considered vendor errors unrecoverable, i.e., something we cannot do
anything about. Recently we discovered that is not the case. We now need
to re-enable them and the only way to do that is to re-enable analysis
of everything. We also catch the mod vendor issues now separately.

Note that a follow-up CL will revert some the changes made here, namely
the one where don't check if the error is recoverable.

Change-Id: Iaf46c5799dd9e61d90588d8dcef23e16d85ad437
Reviewed-on: https://go-review.googlesource.com/c/pkgsite-metrics/+/608575
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Jonathan Amsterdam <jba@google.com>